+++ programs/drift/src/math/orders.rs
@@ -10,7 +10,7 @@ use crate::math::amm::calculate_amm_available_liquidity;
 use crate::math::auction::{is_amm_available_liquidity_source, is_auction_complete};
 use crate::math::casting::Cast;
 use crate::{
-    math, PostOnlyParam, State, BASE_PRECISION_I128, OPEN_ORDER_MARGIN_REQUIREMENT,
+    load, math, PostOnlyParam, State, BASE_PRECISION_I128, OPEN_ORDER_MARGIN_REQUIREMENT,
     PERCENTAGE_PRECISION, PERCENTAGE_PRECISION_U64, PRICE_PRECISION_I128, QUOTE_PRECISION_I128,
     SPOT_WEIGHT_PRECISION,
 };
@@ -25,6 +25,7 @@ use crate::math::spot_balance::{get_strict_token_value, get_token_value};
 use crate::math::spot_withdraw::get_max_withdraw_for_market_with_token_amount;
 use crate::math_error;
 use crate::print_error;
+use crate::state::oracle::OraclePriceData;
 use crate::state::oracle_map::OracleMap;
 use crate::state::perp_market::{PerpMarket, AMM};
 use crate::state::perp_market_map::PerpMarketMap;
@@ -33,6 +34,7 @@ use crate::state::spot_market_map::SpotMarketMap;
 use crate::state::user::{
     MarketType, Order, OrderStatus, OrderTriggerCondition, PerpPosition, User,
 };
+use crate::state::user_map::UserMap;
 use crate::validate;
 
 #[cfg(test)]
@@ -1186,4 +1188,121 @@ fn calculate_free_collateral_delta_for_spot(
             .get_liability_weight(worst_case_token_amount, &MarginRequirementType::Initial)?
             .sub(SPOT_WEIGHT_PRECISION)
     })
+}
+
+#[derive(Eq, PartialEq, Debug)]
+pub struct Level {
+    pub price: u64,
+    pub base_asset_amount: u64,
+}
+
+pub type Side = Vec<Level>;
+
+pub fn find_bids_and_asks_from_users(
+    perp_market: &PerpMarket,
+    oracle_price_date: &OraclePriceData,
+    users: &UserMap,
+    slot: u64,
+    now: i64,
+) -> DriftResult<(Side, Side)> {
+    let mut bids: Side = Vec::with_capacity(8);
+    let mut asks: Side = Vec::with_capacity(8);
+
+    let market_index = perp_market.market_index;
+    let tick_size = perp_market.amm.order_tick_size;
+    let oracle_price = Some(oracle_price_date.price);
+
+    let mut insert_order = |base_asset_amount: u64, price: u64, direction: PositionDirection| {
+        let orders = match direction {
+            PositionDirection::Long => &mut bids,
+            PositionDirection::Short => &mut asks,
+        };
+        let index = match orders.binary_search_by(|level| match direction {
+            PositionDirection::Long => price.cmp(&level.price),
+            PositionDirection::Short => level.price.cmp(&price),
+        }) {
+            Ok(index) => index,
+            Err(index) => index,
+        };
+
+        if index < orders.capacity() {
+            if orders.len() == orders.capacity() {
+                orders.pop();
+            }
+
+            orders.insert(
+                index,
+                Level {
+                    price,
+                    base_asset_amount,
+                },
+            );
+        }
+    };
+
+    for account_loader in users.0.values() {
+        let user = load!(account_loader)?;
+
+        for (_, order) in user.orders.iter().enumerate() {
+            if order.status != OrderStatus::Open {
+                continue;
+            }
+
+            if order.market_type != MarketType::Perp || order.market_index != market_index {
+                continue;
+            }
+
+            // if order is not limit order or must be triggered and not triggered, skip
+            if !order.is_limit_order() || (order.must_be_triggered() && !order.triggered()) {
+                continue;
+            }
+
+            if !order.is_resting_limit_order(slot)? {
+                continue;
+            }
+
+            if now > order.max_ts && order.max_ts != 0 {
+                continue;
+            }
+
+            let existing_position = user.get_perp_position(market_index)?.base_asset_amount;
+            let base_amount = order.get_base_asset_amount_unfilled(Some(existing_position))?;
+            let limit_price = order.force_get_limit_price(oracle_price, None, slot, tick_size)?;
+
+            insert_order(base_amount, limit_price, order.direction);
+        }
+    }
+
+    Ok((bids, asks))
+}
+
+pub fn estimate_price_from_side(side: &Side, depth: u64) -> DriftResult<Option<u64>> {
+    let mut depth_remaining = depth;
+    let mut cumulative_base = 0_u64;
+    let mut cumulative_quote = 0_u128;
+
+    for level in side {
+        let base_delta = level.base_asset_amount.min(depth_remaining);
+        let quote_delta = level.price.cast::<u128>()?.safe_mul(base_delta.cast()?)?;
+
+        cumulative_base = cumulative_base.safe_add(base_delta)?;
+        depth_remaining = depth_remaining.safe_sub(base_delta)?;
+        cumulative_quote = cumulative_quote.safe_add(quote_delta)?;
+
+        if depth_remaining == 0 {
+            break;
+        }
+    }
+
+    let price = if depth_remaining == 0 {
+        Some(
+            cumulative_quote
+                .safe_div(cumulative_base.cast()?)?
+                .cast::<u64>()?,
+        )
+    } else {
+        None
+    };
+
+    Ok(price)
 }