+++ programs/drift/src/validation/order.rs
@@ -6,7 +6,6 @@ use crate::error::{DriftResult, ErrorCode};
 use crate::math::casting::Cast;
 use crate::math::orders::{
     calculate_base_asset_amount_to_fill_up_to_limit_price, is_multiple_of_step_size,
-    order_breaches_oracle_price_bands,
 };
 use crate::state::perp_market::PerpMarket;
 use crate::state::user::{Order, OrderTriggerCondition, OrderType};
@@ -182,19 +181,6 @@ fn validate_limit_order(
         )?;
 
         validate_post_only_order(order, market, valid_oracle_price, slot)?;
-
-        let order_breaches_oracle_price_limits = order_breaches_oracle_price_bands(
-            order,
-            valid_oracle_price.ok_or(ErrorCode::InvalidOracle)?,
-            slot,
-            market.amm.order_tick_size,
-            market.margin_ratio_initial,
-            market.margin_ratio_maintenance,
-        )?;
-
-        if order_breaches_oracle_price_limits {
-            return Err(ErrorCode::OrderBreachesOraclePriceLimits);
-        }
     }
 
     validate_limit_order_auction_params(order)?;
@@ -436,28 +422,10 @@ fn validate_auction_params(order: &Order) -> DriftResult {
     Ok(())
 }
 
-pub fn validate_spot_order(
-    order: &Order,
-    valid_oracle_price: Option<i64>,
-    slot: u64,
-    step_size: u64,
-    tick_size: u64,
-    margin_ratio_initial: u32,
-    margin_ratio_maintenance: u32,
-    min_order_size: u64,
-) -> DriftResult {
+pub fn validate_spot_order(order: &Order, step_size: u64, min_order_size: u64) -> DriftResult {
     match order.order_type {
         OrderType::Market => validate_market_order(order, step_size, min_order_size)?,
-        OrderType::Limit => validate_spot_limit_order(
-            order,
-            valid_oracle_price,
-            slot,
-            step_size,
-            tick_size,
-            min_order_size,
-            margin_ratio_initial,
-            margin_ratio_maintenance,
-        )?,
+        OrderType::Limit => validate_spot_limit_order(order, step_size, min_order_size)?,
         OrderType::TriggerMarket => {
             validate_trigger_market_order(order, step_size, min_order_size)?
         }
@@ -468,16 +436,7 @@ pub fn validate_spot_order(
     Ok(())
 }
 
-fn validate_spot_limit_order(
-    order: &Order,
-    valid_oracle_price: Option<i64>,
-    slot: u64,
-    step_size: u64,
-    tick_size: u64,
-    min_order_size: u64,
-    margin_ratio_initial: u32,
-    margin_ratio_maintenance: u32,
-) -> DriftResult {
+fn validate_spot_limit_order(order: &Order, step_size: u64, min_order_size: u64) -> DriftResult {
     validate_base_asset_amount(order, step_size, min_order_size, order.reduce_only)?;
 
     if order.price == 0 && !order.has_oracle_price_offset() {
@@ -501,19 +460,6 @@ fn validate_spot_limit_order(
             ErrorCode::InvalidOrder,
             "post only limit order cant have auction"
         )?;
-
-        let order_breaches_oracle_price_limits = order_breaches_oracle_price_bands(
-            order,
-            valid_oracle_price.ok_or(ErrorCode::InvalidOracle)?,
-            slot,
-            tick_size,
-            margin_ratio_initial,
-            margin_ratio_maintenance,
-        )?;
-
-        if order_breaches_oracle_price_limits {
-            return Err(ErrorCode::OrderBreachesOraclePriceLimits);
-        }
     }
 
     validate_limit_order_auction_params(order)?;